refactor: create additional folder for better commands structure
